+++ Assets/Runtime/Core/AsyncLoad/ILoadPresenter.cs
@@ -1,4 +1,4 @@
-namespace Data.Provider
+namespace Data.AsyncLoad
 {
     public interface ILoadPresenter
     {
