+++ Assets/Runtime/Core/AsyncLoad/ReactiveCollection.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace Data.Provider
+namespace Data.AsyncLoad
 {
     public class ReactiveCollection<TKey, TValue> : IReactiveCollection<TValue>
     {
