+++ Assets/Runtime/Core/Enchantment/EnchanceTable/EnchantmentTablePresenter.cs
@@ -1,5 +1,5 @@
 using Data.Model.Item;
-using Data.Provider;
+using Data.AsyncLoad;
 using Data.References;
 using Enchantment.Item;
